Updated WorldMap to use connections
Started route work in Train, now every MapObj (Station or Junction) has
coordinates ready for them to be added to the map. I created a mapblip
image which could be overlayed upon clicking a mapobj.

Every MapObj has an array of connections, which includes the length as a
weighting (calculated using mapobj coordinates)