TrainUpgrade classes and ArrayList + new WildCards
There is now a way to add and remove upgrades from a train - which are
stored in an ArrayList<TrainUpgrade> upgrades in the Train class - tried
to implement teleport train and GoFasterStripes but need more UI stuff
to actually test it :)

The shop needs to use the CardFactory and TrainUpgrade instances need to
be made upon purchasing the said upgrade.

needs work.